refactor: rework dashboard
